General code refactoring
Just some quality of life code changes which should not
influence the behaviour of the program.

The intention is to improve readability.

+++ utils.c
@@ -80,9 +80,9 @@ void strcpy_nospaces(char *dest, char *src) {
  * AC_SEARCH_LIBS(pow, [c m], AC_DEFINE([HAVE_POW]))
  */
 
-int gauge_to_si(u64 bignum, char **str) {
+int gauge_to_si(unsigned long long bignum, char **str) {
 	long unsigned int i = 0;
-	u64 tmpll;
+	unsigned long long tmpll;
 	static char units[] = "kMGTPE";
 
 	tmpll = bignum;
